MQTT Proof of concept
The goals where to be able to send logs, unlock/deny, and health check
events to a MQTT broker. Also being able process a new card list by
subscribing to the access_list topic. This allows for updating the
cards.txt file without having to plug a laptop into the door controller.

This change also includes all the plumbing for setting up the ETH
connection and getting the MQTT broker IP using a DNS lookup.

It also provides a way to change the default configuration.

This is a POC which means it's not quite ready for prime time. It is
however ready for more real world testing.