QoS and topic structure changes
Use the client id as part of the publish topic. This will make it
possible to know what client the messsage came from. Otherwise, the
client id would have to be placed into every payload

Update QoS from 2 to 1. 2 is pretty chatty and 1 still provides the
reliability we want. Also updated the code to handle duplicates which
are not possible when using QoS 1 over 2

Set retain flag to false as the messages shouldn't be retained. If a
client disconnects, then it should receive messages when it reconnects.
That is assuming the client is using a persistent session.